requirements.DEVELOPMENT.txt
-r requirements.LOCAL.txt
-r requirements.txt
Werkzeug==3.0.1
black==23.7.0
coverage==7.4.0
django-constance[database]==3.1.0
django-fakeredis==0.1.2
django-sslserver==0.22
jsonschema==4.17.3
pre-commit==3.3.3
ruff==0.0.264
pytest==8.2.1
pytest-django==4.8.0
pytest-playwright==0.5.0